Updated buy method
Removed the buy now option, users now have to add items to the cart, regardless of whether they want to buy just one product.

Buy now fixed to receive a list of products (cart)

Add to cart implemented, when a user add's item to the cart creates a collection cart with a list of products and their quantity

**FIREBASE**

Updated request collection, now contains:
- Address
- If the user pin points his localization via maps, it will save the Latitude and Longitude
- NIF
- Observations
- Product Name
- Product Price
(All taken from collection product)
- Status - Starts at 'In Progress' and it will updated only by the restaurant people
- Timestamp
- userId , so it becomes easier to get all the requests for each user (taken from user collection)

Add cart collection which contains:
- Document named with UserId that has a list of products (contains all the fields from product) and his quantity

Every time that a new user add something to the cart, it creates a new document with the userId. It becomes easier to now which cart belongs to each user.

Next implementations:
- Button to see what requests the user have done
- Credit Card payment API